feat: enhance endorsements page
