+++ src/app/endorsements/page.tsx
@@ -1,9 +1,10 @@
 import type { Metadata } from 'next'
 
-import Endorsements from '@/components/endorsements'
 import PageHeader from '@/components/page-header'
 import { getMetadata } from '@/lib/metadata'
 import { getEndorsements } from '@/services/endorsements'
+
+import Endorsements from './endorsements'
 
 export const metadata: Metadata = getMetadata({
   title: 'Endorsements',
