+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './use-available-devices'
-export * from './use-endorsements'
 export * from './use-insights'
 export * from './use-isomorphic-layout-effect'
 export * from './use-media-query'
