--- src/types/users.ts
@@ -1,5 +1,0 @@
-export type User = {
-  id: string
-  name: string
-  image: string
-}
